+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Adds --require switch for user configuration
 * Deprecates --compass-with-config config.rb in favor of --compass --require config.rb
+* Ignores @import statements. Users will need to run csscss on those directly
 
 ## 1.2.0 - 4/14/2013 ##